Code refactor
- codice migliorato
- rimossi import inutilizzati
- separato modulo per gestione S3
- auth0 viene chiamato con token ricevuto in input, risparmiando la generazione di un token nuovo
