+++ .env.example
@@ -21,6 +21,6 @@ BUCKET_NAME=BUCKET
 
 # Auth0
 AUTH0_DOMAIN=AUTH0_DOMAIN
-AUTH0_SECRET=AUTH0_SECRET
+# AUTH0_SECRET=AUTH0_SECRET
 AUTH0_API_AUDIENCE=AUTH0_API_AUDIENCE
 AUTH0_ISSUER=AUTH0_ISSUER
